tests/e2e/python37/requirements.txt
azure-functions
pyodbc==4.0.30